django/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.2.2               # via kombu
apipkg==1.4               # via execnet
awesome-slugify==1.6.5
awscli==1.11.160
babel==2.5.1              # via flower
beautifulsoup4==4.6.0
billiard==3.5.0.3         # via celery
boto==2.48.0
botocore==1.7.18          # via awscli, s3transfer
cached-property==1.3.1    # via django-url-filter
celery==4.1.0
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
colorama==0.3.7           # via awscli
coreapi==2.3.1            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
decorator==4.1.2          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid
diff-match-patch==20121119
django-allauth==0.33.0
django-autocomplete-light==2.3.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.9.1
django-filter==1.0.4
django-guardian==1.4.9
django-model-utils==3.0.0
django-redis-cache==1.7.1
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-storages==1.6.5
django-url-filter==0.3.4
django-webpack-loader==0.4.1
django==1.11
djangorestframework==3.6.4
docopt==0.6.2             # via ptpython
docutils==0.14            # via awscli, botocore
enum-compat==0.0.2        # via django-url-filter
execnet==1.4.1            # via pytest-xdist
faker==0.8.4
flower==0.9.2
fuzzywuzzy==0.15.1
hiredis==0.2.0
idna==2.6                 # via requests
imagehash==3.7
ipdb==0.10.3
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.0
itypes==1.1.0             # via coreapi
jedi==0.11.0              # via ipython, ptpython
jinja2==2.9.6             # via coreschema
jmespath==0.9.3           # via botocore
kombu==4.1.0              # via celery
markdown==2.6.9
markupsafe==1.0           # via jinja2
numpy==1.13.1             # via imagehash, pywavelets, scipy
oauthlib==2.0.4           # via requests-oauthlib
olefile==0.44             # via pillow
openapi-codec==1.3.2      # via django-rest-swagger
parso==0.1.0              # via jedi
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
pillow==4.2.1
prompt-toolkit==1.0.15    # via ipython, ptpython
psycopg2==2.7.3.1
ptpython==0.41
ptyprocess==0.5.2         # via pexpect
py==1.4.34                # via pytest
pyasn1==0.3.6             # via rsa
pygments==2.2.0           # via ipython, ptpython
pypdf2==1.26.0
pytest-django==3.1.2
pytest-forked==0.2        # via pytest-xdist
pytest-xdist==1.20.0
pytest==3.2.2
python-dateutil==2.6.1    # via botocore, faker
python-levenshtein==0.12.0
python3-openid==3.1.0     # via django-allauth
pytz==2017.2              # via babel, celery, django, flower
pywavelets==0.5.2         # via imagehash
pyyaml==3.12              # via awscli
raven==6.2.1
redis==2.10.6
regex==2017.9.23          # via awesome-slugify
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4
rsa==3.4.2                # via awscli
s3transfer==0.1.11        # via awscli
scipy==0.19.1             # via imagehash
selenium==3.6.0
simplegeneric==0.8.1      # via ipython
simplejson==3.11.1        # via django-rest-swagger
six==1.11.0               # via django-autocomplete-light, django-extensions, django-guardian, django-rest-auth, django-url-filter, faker, imagehash, prompt-toolkit, python-dateutil, traitlets
sorl-thumbnail==12.4a1
sqlparse==0.2.4           # via django-debug-toolbar
tornado==4.5.2            # via flower
traitlets==4.3.2          # via ipython
unidecode==0.4.21         # via awesome-slugify
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
uwsgi==2.0.15
vine==1.1.4               # via amqp
wand==0.4.4
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.12.2
